Encoding fixes for prompt studio k8 changes

+++ backend/prompt_studio/prompt_studio_core_v2/models.py
@@ -7,13 +7,19 @@ from account_v2.models import User
 from adapter_processor_v2.models import AdapterInstance
 from django.db import models
 from django.db.models import QuerySet
+from file_management.file_management_helper import FileManagerHelper
 from prompt_studio.prompt_studio_core_v2.constants import DefaultPrompts
+from utils.constants import FeatureFlag
+from utils.file_storage.constants import FileStorageType
+from utils.file_storage.helpers.common_file_helper import FileStorageHelper
 from utils.file_storage.helpers.prompt_studio_file_helper import PromptStudioFileHelper
 from utils.models.base_model import BaseModel
 from utils.models.organization_mixin import (
     DefaultOrganizationManagerMixin,
     DefaultOrganizationMixin,
 )
+
+from unstract.flags.feature_flag import check_feature_flag_status
 
 logger = logging.getLogger(__name__)
 
@@ -133,20 +139,37 @@ class CustomTool(DefaultOrganizationMixin, BaseModel):
 
     def delete(self, organization_id=None, *args, **kwargs):
         # Delete the documents associated with the tool
-        file_path = PromptStudioFileHelper.handle_sub_directory_for_prompt_studio(
-            organization_id,
-            is_create=False,
-            user_id=self.created_by.user_id,
-            tool_id=str(self.tool_id),
-        )
-        if organization_id:
+        if check_feature_flag_status(FeatureFlag.REMOTE_FILE_STORAGE):
+            file_path = FileManagerHelper.handle_sub_directory_for_prompt_studio(
+                organization_id,
+                is_create=False,
+                user_id=self.created_by.user_id,
+                tool_id=str(self.tool_id),
+            )
+            if organization_id:
+                try:
+                    shutil.rmtree(file_path)
+                except FileNotFoundError:
+                    logger.error(f"The folder {file_path} does not exist.")
+                except OSError as e:
+                    logger.error(f"Error: {file_path} : {e.strerror}")
+                    # Continue with the deletion of the tool
+        else:
+            fs_instance = FileStorageHelper.initialize_file_storage(
+                type=FileStorageType.PERMANENT
+            )
+            file_path = PromptStudioFileHelper.get_or_create_prompt_studio_subdirectory(
+                organization_id,
+                is_create=False,
+                user_id=self.created_by.user_id,
+                tool_id=str(self.tool_id),
+            )
             try:
-                shutil.rmtree(file_path)
+                fs_instance.rm(file_path, True)
             except FileNotFoundError:
-                logger.error(f"The folder {file_path} does not exist.")
-            except OSError as e:
-                logger.error(f"Error: {file_path} : {e.strerror}")
-                # Continue with the deletion of the tool
+                # Supressed to handle cases when the remote
+                # file is missing or already deleted
+                pass
         super().delete(*args, **kwargs)
 
     class Meta:
